refactor: CardTitleWrapper CardTitleContainer 컴포넌트로 변경
